Dodan button logout i zapoceto logout
