+++ package.json
@@ -13,6 +13,7 @@
     "express": "^4.17.1",
     "hbs": "^4.1.1",
     "mysql": "^2.18.1",
+    "node-fetch": "^2.6.1",
     "path": "^0.12.7"
   }
 }
