+++ templates/views/mainpage.hbs
@@ -15,6 +15,7 @@
     </div>
     <textarea id="textAreaDescription" cols="30" rows="10"></textarea>
     <button id="addTaskButton">Add task</button>
+    <button id="logoutButton">Log out</button>
     <script src="./js/index.js"></script>
 </body>
